+++ core-ui/src/components/Predefined/Create/index.js
@@ -14,5 +14,5 @@ export * from './Certificates/Certificates.create.js';
 export * from './DNSEntries/DNSEntries.create.js';
 export * from './Functions/Functions.create.js';
 export * from './Roles/Roles.create.js';
-export * from './Jobs/CronJobs/CronJobs.create.js';
-export * from './Jobs/Jobs/Jobs.create.js';
+export * from './Jobs/CronJobs.create.js';
+export * from './Jobs/Jobs.create.js';
